LPD-25046 Move WebDAVEnvironmentConfigClassTestRule to test/rule
